data_logger/requirements.logger.txt
requests
xmltodict
cryptography
python-dotenv